modified the timer in hc_sr, and make the beeper ringing.

+++ inc/hc_sr04.h
@@ -3,7 +3,7 @@
 
 #include "stm32l476xx.h"
 
-void wait(int TIME);
+void wait(float TIME);
 void send_trigger();
 double get_distance();
